refactor: 移除converter转成内部类、read等调整
