+++ spring-extension-context/src/main/java/com/livk/context/fastexcel/resolver/ExcelMethodReturnValueHandler.java
@@ -19,13 +19,19 @@ package com.livk.context.fastexcel.resolver;
 import com.livk.commons.util.AnnotationUtils;
 import com.livk.context.fastexcel.FastExcelSupport;
 import com.livk.context.fastexcel.annotation.ResponseExcel;
-import com.livk.context.fastexcel.converter.ExcelHttpMessageConverter;
 import com.livk.context.fastexcel.exception.ExcelExportException;
 import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
+import org.springframework.core.ResolvableType;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpInputMessage;
+import org.springframework.http.HttpOutputMessage;
 import org.springframework.http.MediaType;
 import org.springframework.http.MediaTypeFactory;
+import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.http.converter.HttpMessageNotWritableException;
 import org.springframework.http.server.ServletServerHttpResponse;
 import org.springframework.lang.NonNull;
 import org.springframework.util.Assert;
@@ -34,6 +40,9 @@ import org.springframework.web.method.support.AsyncHandlerMethodReturnValueHandl
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 import java.io.IOException;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 
 /**
  * <p>
@@ -62,15 +71,15 @@ public class ExcelMethodReturnValueHandler implements AsyncHandlerMethodReturnVa
 		ResponseExcel responseExcel = AnnotationUtils.getAnnotationElement(returnType, ResponseExcel.class);
 		Assert.notNull(response, "response not be null");
 		Assert.notNull(responseExcel, "responseExcel not be null");
-		ExcelHttpMessageConverter converter = ExcelHttpMessageConverter.writeExcel(returnType, responseExcel);
-		if (converter.canWrite(returnType.getParameterType(), MediaType.ALL)) {
-			String fileName = FastExcelSupport.fileName(responseExcel);
+		ExcelHttpMessageWriter writer = new ExcelHttpMessageWriter(returnType, responseExcel);
+		if (writer.canWrite(returnType.getParameterType(), MediaType.ALL)) {
+			String fileName = ResponseExcel.Utils.parseName(responseExcel);
 			String contentType = MediaTypeFactory.getMediaType(fileName)
 				.map(MediaType::toString)
 				.orElse("application/vnd.ms-excel");
 			response.setCharacterEncoding(UTF8);
 			response.setHeader(HttpHeaders.CONTENT_DISPOSITION, "attachment;filename=" + fileName);
-			converter.write(returnValue, MediaType.valueOf(contentType), new ServletServerHttpResponse(response));
+			writer.write(returnValue, MediaType.valueOf(contentType), new ServletServerHttpResponse(response));
 		}
 		else {
 			throw new ExcelExportException("the return class is not java.util.Collection or java.util.Map");
@@ -80,6 +89,54 @@ public class ExcelMethodReturnValueHandler implements AsyncHandlerMethodReturnVa
 	@Override
 	public boolean isAsyncReturnValue(Object returnValue, @NonNull MethodParameter returnType) {
 		return AnnotationUtils.hasAnnotationElement(returnType, ResponseExcel.class);
+	}
+
+	@RequiredArgsConstructor
+	private static class ExcelHttpMessageWriter implements HttpMessageConverter<Object> {
+
+		private final MethodParameter parameter;
+
+		private final ResponseExcel responseExcel;
+
+		@Override
+		public boolean canRead(@NonNull Class<?> clazz, MediaType mediaType) {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public boolean canWrite(@NonNull Class<?> type, MediaType mediaType) {
+			return Collection.class.isAssignableFrom(type) || Map.class.isAssignableFrom(type);
+		}
+
+		@NonNull
+		@Override
+		public List<MediaType> getSupportedMediaTypes() {
+			return List.of();
+		}
+
+		@NonNull
+		@Override
+		public Object read(@NonNull Class<?> parameterType, HttpInputMessage inputMessage)
+				throws HttpMessageNotReadableException {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public void write(@NonNull Object returnValue, MediaType contentType, @NonNull HttpOutputMessage outputMessage)
+				throws IOException, HttpMessageNotWritableException {
+			if (returnValue instanceof Collection) {
+				Class<?> excelModelClass = ResolvableType.forMethodParameter(parameter).resolveGeneric(0);
+				FastExcelSupport.write(outputMessage.getBody(), excelModelClass, responseExcel.template(),
+						Map.of("sheet", (Collection<?>) returnValue));
+			}
+			else if (returnValue instanceof Map) {
+				@SuppressWarnings("unchecked")
+				Map<String, Collection<?>> result = (Map<String, Collection<?>>) returnValue;
+				Class<?> excelModelClass = ResolvableType.forMethodParameter(parameter).getGeneric(1).resolveGeneric(0);
+				FastExcelSupport.write(outputMessage.getBody(), excelModelClass, responseExcel.template(), result);
+			}
+		}
+
 	}
 
 }
